requirements.txt
nltk>=3.3
pandas>=0.23.3
soundfile>=0.10.2
sox>=1.3.3
absl-py>=0
numpy~=1.19.5
tensorflow~=2.4.1
#tf-models-official~=2.4.0
